requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.1.0
audioread==3.0.1
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.12.3
billiard==4.2.0
blis==0.7.11
bootstrap4==0.1.0
cachetools==5.5.2
catalogue==2.0.10
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.18.1
confection==0.1.5
constantly==23.10.4
cryptography==44.0.1
cymem==2.0.8
daphne==4.1.2
decorator==5.1.1
distlib==0.3.8
distro==1.9.0
Django==4.2.13
django-appconf==1.1.0
django-bootstrap4==24.3
django-celery-email==3.0.0
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.3.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
exceptiongroup==1.2.2
factory_boy==3.0.1
Faker==26.0.0
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
google-api-core==2.24.1
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.24.5
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
inflection==0.5.1
iniconfig==2.0.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.3.7
langcodes==3.4.0
language_data==1.2.0
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
murmurhash==1.0.10
networkx==3.2.1
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
packaging==24.1
pandas==2.2.2
pbr==6.0.0
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
pooch==1.8.2
preshed==3.0.9
prompt_toolkit==3.0.47
proto-plus==1.26.0
protobuf==5.29.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyparsing==3.2.1
pytest==8.3.2
pytest-django==4.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==5.3.1
redis==5.0.8
referencing==0.36.2
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.23.1
rsa==4.9
service-identity==24.2.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.5
soxr==0.4.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.0
srsly==2.4.8
stevedore==5.2.0
sympy==1.13.1
thinc==8.2.5
threadpoolctl==3.5.0
tokenizers==0.19.1
tomli==2.0.1
tqdm==4.66.4
Twisted==24.11.0
txaio==23.1.1
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==3.0.1
urllib3==2.2.2
vine==5.1.0
virtualenv==20.26.3
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
wrapt==1.16.0
yarl==1.9.4
zope.interface==7.2
